// akadon-frontend/src/app/models/student.ts
export class Student {
  studentId?: number;
  email?: string;
  password?: string;
  fullName?: string;
  gender?: Boolean;
  city?: string;
  district?: string;
  ward?: string;
  phonenumber?: string;
  birthday?: Date;
  role?: string;
  status?: Boolean;
  image?: string;
  learntCourseNumber?: number;
  lastLogin?: Date;
  rating?: number;
  behaviorPoint?: number;
  verificationCode?: string;
  otpRequestTime?: Date;
}